Use default data from /data folder
